Moving People: January 2000–July 2021
BTS is withholding the scheduled release of the passenger and combined indexes for January. The passenger index is a statistical estimate of airline passenger travel and other components based on historical trends up to December 2019. As a result, the estimates have yet to fully account for the impact of the coronavirus. Air freight is also a statistical estimate. Since air freight makes up a smaller part of the freight index, the freight TSI is being released as scheduled.



Description



Statisticians use the process of seasonal-adjustment to uncover trends in data. Monthly data, for instance, are influenced by the number of days and the number of weekends in a month as well as by the timing of holidays and seasonal activity. These influences make it difficult to see underlying changes in the data. Statisticians use seasonal adjustment to control for these influences.



Controlling of seasonal influences allows measurement of real monthly changes; short and long term patterns of growth or decline; and turning points. Data for one month can be compared to data for any other month in the series and the data series can be ranked to find high and low points. Any observed differences are “real” differences; that is, they are differences brought about by changes in the data and not brought about by a change in the number of days or weekends in the month, the occurrence or non-occurrence of a holiday, or seasonal activity.